Force string comparison for id field in patch.
The patch code expected the id in the data object to be a string, while Flask-REST-JSONAPI itself delivers it as integer if it is one in the schema.

This made it a bit awkward to transform a loaded object into a patch object. This commit forces casting both sides of the comparison to string.