Fix using multiple includes with sub-schemas, fixes #56 and #29.
Due to the layout of the loop creating the related_schema isntances, all
but the last related_include were lost. They are thus now collected in
the inital step where all include_paths are evaluated, so that the
schema creation code can use the colelcted sub-includes later.

It is still not perfect, because the related_schema is created multiple
times and later discarded in some cases, but that's a minor issue and
was there before as well. The important thing is that the last schema
created is complete now.