schema.py: Fix for get_nested_fields
Having a list field on schema is giving an error after upgrading to the latest version. The get_nested_fields function is trying to access an attribute 'container' on a list field, which as per latest marshmallow version does not exist. So replacing it with a new field called 'inner'.